rename SshApiService, XmlRpcService and JavaClient classes, move JavaClient to test package
